requirements.txt
pytest
beautifulsoup4
mysql-connector
